requirements.txt
pydub==0.25.1
transformers
openai-whisper
sentencepiece